+++ app/Http/Controllers/InvoiceSettingsClientDetailsController.php
@@ -24,7 +24,7 @@ class InvoiceSettingsClientDetailsController extends Controller{
 	
 	public function show(Request $request) : mixed{
 		
-		$company_id = Sanitize::input($request->input('company_id'));
+		$company_id = (int) Sanitize::input($request->input('company_id'));
 		
 		return (new SettingsArrangedFields($this->getClientDetailFields(), $request, $company_id))->fetchArrangedFieldsData(ClientsCustomField::class);
 
@@ -33,7 +33,7 @@ class InvoiceSettingsClientDetailsController extends Controller{
 	
 	public function saveOrUpdate(Request $request) : mixed{
 
-		$company_id = Sanitize::input($request->input('company_id'));
+		$company_id = (int) Sanitize::input($request->input('company_id'));
 
 		$settings_arranged_fields = new SettingsArrangedFields($this->getClientDetailFields(), $request, $company_id);
